+++ packages/plugin-packages/mashroom-portal/src/backend/controllers/PortalAppController.ts
@@ -250,6 +250,15 @@ export default class PortalAppController {
         }
 
         const resourceUri = `${portalApp.resourcesRootUri}/${resourcePath}`;
+
+        // Security check: A proxy target could be a sub path of the resource base URL, so,
+        //  make sure this route is not misused to access an API endpoint
+        if (portalApp.proxies && Object.values(portalApp.proxies).some(({targetUri}) => resourceUri.startsWith(targetUri))) {
+            logger.error('Attempted access to an API endpoint via resource request:', resourceUri);
+            res.sendStatus(401);
+            return false;
+        }
+
         logger.debug(`Sending portal app resource: ${resourceUri}`);
 
         try {